add by guansq at 2011-9-9 更改考核指标分数允许录入小数

+++ 06_Office_HR/src/com/xpsoft/oa/action/kpi/HrPaPerformanceindexAction.java
@@ -331,7 +331,7 @@ public class HrPaPerformanceindexAction extends BaseAction {
 					}
 					indexScore.setPi(piNew);
 					indexScore.setPisType(piNew.getMode());
-					indexScore.setPisScore(BigDecimal.valueOf(Long.parseLong(itemArray[1])));
+					indexScore.setPisScore(BigDecimal.valueOf(Float.parseFloat(itemArray[1])));
 					indexScore.setPisDesc(itemArray[2]);
 					indexScoreList.add(indexScore);
 					//定量考核添加计算公式
